backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
bcrypt==4.2.0
passlib==1.7.4
pydantic==2.9.2
typing_extensions==4.12.2
motor==3.6.0
pymongo==4.9.2
spacy==3.8.2
thinc==8.3.2
pytesseract==0.3.13
pdf2image==1.17.0
python-jose==3.3.0
cryptography==43.0.1
requests==2.32.3
certifi==2024.8.30
idna==3.10
urllib3==2.2.3
